FIX: Handle DICOMs not in Archives [GEAR-381]

+++ run.py
@@ -10,8 +10,12 @@ import pprint
 import zipfile
 import pydicom
 import logging
-import flywheel
+import tempfile
+from pathlib import Path
 from pprint import pprint as pp
+
+import flywheel
+
 from dicom_metadata import dicom_header_extract
 from util import quote_numeric_string, ensure_filename_safety
 
@@ -74,13 +78,13 @@ def _archive_session(fw, session, archive_project):
     session.update({'subject': {'_id': subject.id}})
 
 
-def _create_archive(content_dir, arcname, zipfilepath=None):
+def _create_archive(content_dir, arcname, file_list, zipfilepath=None):
     """Create zip archive from content_dir"""
     if not zipfilepath:
         zipfilepath = content_dir + '.zip'
     with zipfile.ZipFile(zipfilepath, 'w', zipfile.ZIP_DEFLATED, allowZip64=True) as zf:
         zf.write(content_dir, arcname)
-        for fn in os.listdir(content_dir):
+        for fn in file_list:
             zf.write(os.path.join(content_dir, fn), os.path.join(os.path.basename(arcname), fn))
     return zipfilepath
 
@@ -89,26 +93,51 @@ def _extract_archive(zip_file_path, extract_location):
     """Extract zipfile to <zip_file_path> and return the path to the directory containing the dicoms,
     which should be the zipfile name without the zip extension."""
     import zipfile
-    if not zipfile.is_zipfile(zip_file_path):
-        log.warning('{} is not a Zip File!'.format(zip_file_path))
-        return None
 
     with zipfile.ZipFile(zip_file_path) as ZF:
-        if '/' in ZF.namelist()[0]:
-            extract_dest = os.path.join(extract_location, ZF.namelist()[0].split('/')[0])
-            ZF.extractall(extract_location)
-            return extract_dest
-        else:
-            extract_dest = os.path.join(extract_location, os.path.basename(zip_file_path).split('.zip')[0])
-            if not os.path.isdir(extract_dest):
-                log.debug('Creating extract directory: {}'.format(extract_dest))
-                os.mkdir(extract_dest)
-            log.debug('Extracting {} archive to: {}'.format(zip_file_path, extract_dest))
-            ZF.extractall(extract_dest)
-            return extract_dest
+
+        extract_dest = os.path.join(extract_location, os.path.basename(zip_file_path).split('.zip')[0])
+        if not os.path.isdir(extract_dest):
+            log.debug('Creating extract directory: {}'.format(extract_dest))
+            os.mkdir(extract_dest)
+        log.debug('Extracting {} archive to: {}'.format(zip_file_path, extract_dest))
+        ZF.extractall(extract_dest)
+    return extract_dest
 
 
-def _export_dicom(dicom_file, acquisition, session, subject, project, config):
+def _retrieve_path_list(file_path):
+    """ For a given DICOM archive, check to see if it's a zip file.  If it is zip, extract the
+        archive, and return a list of all the files in the archive.  If it's not a zip,
+        simply return the file. returns a tuple with a list of fille paths and a boolean to indicate
+        if the file was a zip or not.
+        
+ 
+        
+        RETURNS: tuple ( [<file_paths>], is_zip )
+        
+    """
+    file_path = Path(file_path)
+    
+    if zipfile.is_zipfile(file_path):
+        zf = zipfile.ZipFile(file_path)
+        is_zip = True
+        zip_list = zf.namelist()
+
+    else:
+        is_zip = False
+        zip_list = [file_path.as_posix()]
+    
+    # Remove any entries from the list that are directories:
+    file_list = []
+    for f in zip_list:
+        if f and f[-1] != '/':
+            file_list.append(f)
+    
+    
+    return((file_list, is_zip))
+        
+
+def _export_dicom(dicom_file, tmp_dir, acquisition, session, subject, project, config):
     """ For a given DICOM archive, or file(?) update the DICOM header metadata according to the
         metadata in Flywheel.
 
@@ -119,9 +148,13 @@ def _export_dicom(dicom_file, acquisition, session, subject, project, config):
     update_keys = []
 
     # Download the dicom archive
-    dicom_file_path = os.path.join('/tmp', dicom_file.name)
-    dicom_file.download(dicom_file_path)
 
+
+    dicom_file_path = os.path.join(tmp_dir, dicom_file.name)
+    dicom_file.download(dicom_file_path)
+    
+    dicom_path_list = _retrieve_path_list(dicom_file_path)
+    
     # For the downloaded file, extract the metadata
     local_dicom_header = dicom_header_extract(dicom_file_path)
     if not local_dicom_header:
@@ -193,13 +226,14 @@ def _export_dicom(dicom_file, acquisition, session, subject, project, config):
 
     # Iterate through the DICOM files and update the values according to the flywheel_dicom_header
     log.info('The following keys will be updated: {}'.format(update_keys))
-    upload_file_path = _modify_dicom_archive(dicom_file_path, update_keys, flywheel_dicom_header)
+    upload_file_path = _modify_dicom_archive(dicom_file_path, update_keys, flywheel_dicom_header, 
+                                             dicom_path_list, tmp_dir)
 
 
     return upload_file_path
 
 
-def _modify_dicom_archive(dicom_file_path, update_keys, flywheel_dicom_header):
+def _modify_dicom_archive(dicom_file_path, update_keys, flywheel_dicom_header, dicom_file_list, tmp_dir):
     """Given a dicom archive <dicom_file_path>, iterate through a list of keys <update_keys> and
     modify each file in the archive with the value in the passed in dict <flywheel_dicom_header>
 
@@ -207,19 +241,24 @@ def _modify_dicom_archive(dicom_file_path, update_keys, flywheel_dicom_header):
 
     """
     import pydicom
-
+    dicom_files, is_zip = dicom_file_list
     # Extract the archive
-    dicom_base_folder = _extract_archive(dicom_file_path, '/tmp')
-
+    if is_zip:
+        dicom_base_folder = _extract_archive(dicom_file_path, tmp_dir)
+    else:
+        dicom_base_folder, base = os.path.split(dicom_file_path)
+        
     # Remove the zipfile
-    os.remove(dicom_file_path)
+    # Still explicitly removing this because we later create a zip archive of the same name
+    if os.path.exists(dicom_file_path) and zipfile.is_zipfile(dicom_file_path):
+        log.debug('Removing zip file {}'.format(dicom_file_path))
+        os.remove(dicom_file_path)
 
-
-    # For each file in the archive, update the keys
-    dicom_files = os.listdir(dicom_base_folder)
     log.info('Updating {} keys in {} dicom files...'.format(len(update_keys), len(dicom_files)))
-    for df in sorted(dicom_files):
+    # for df in sorted(dicom_files):
+    for df in dicom_files:
         dfp = os.path.join(dicom_base_folder, df)
+        log.debug('Reading {}'.format(dfp))
         try:
             dicom = pydicom.read_file(dfp, force=False)
         except:
@@ -245,8 +284,14 @@ def _modify_dicom_archive(dicom_file_path, update_keys, flywheel_dicom_header):
 
 
     # Package up the archive
-    log.debug('Packaging archive: {}'.format(dicom_base_folder))
-    modified_dicom_file_path = _create_archive(dicom_base_folder, os.path.basename(dicom_base_folder))
+
+    if is_zip:
+        log.debug('Packaging archive: {}'.format(dicom_base_folder))
+        modified_dicom_file_path = _create_archive(dicom_base_folder,
+                                                   os.path.basename(dicom_base_folder),
+                                                   dicom_files)
+    else:
+        modified_dicom_file_path = dicom_file_path
 
     return modified_dicom_file_path
 
@@ -272,54 +317,54 @@ def _export_files(fw, acquisition, export_acquisition, session, subject, project
                                             session.label,
                                             acquisition.label,
                                             f.name))
-        if f.type == 'dicom':
-            upload_file_path = _export_dicom(f, acquisition, session, subject, project, config)
-        else:
-            upload_file_path = os.path.join('/tmp', f.name)
-            f.download(upload_file_path)
-
-        # Upload the file to the export_acquisition
-        log.debug("Uploading %s to %s" % (f.name, export_acquisition.label))
-
-        # Add logic around retrying failed uploads
-        max_attempts = 5
-        attempt = 0
-        while attempt < max_attempts:
-            attempt +=1
-            s = export_acquisition.upload_file(upload_file_path)
-            log.info('Upload status = {}'.format(s))
-            export_acquisition = fw.get_acquisition(export_acquisition.id)
-            file_names = [ x.name for x in export_acquisition.files ]
-            log.debug(file_names)
-            if os.path.basename(upload_file_path) not in file_names:
-                log.warning('Upload failed for {} - retrying...'.format(os.path.basename(upload_file_path)))
+        
+        with tempfile.TemporaryDirectory() as temp_dir:
+            
+            if f.type == 'dicom':
+                upload_file_path = _export_dicom(f, temp_dir, acquisition, session, subject, project, config)
             else:
-                log.info('Successfully exported: {}'.format(os.path.basename(upload_file_path)))
-                break
-
-        # Delete the uploaded file locally.
-        log.debug('Removing local file: %s' % (upload_file_path))
-        os.remove(upload_file_path)
-
-        # Update file metadata
-        if f.modality:
-            log.debug('Updating modality to %s for %s' % (f.modality, f.name))
-            export_acquisition.update_file(f.name, modality=f.modality)
-        if not f.modality and f.name.endswith('mriqc.qa.html'):
-            # Special case - mriqc output files do not have modality set, so
-            # we must set the modality prior to the classification to avoid errors.
-            export_acquisition.update_file(f.name, modality='MR')
-        if f.type:
-            log.debug('Updating type to %s for %s' % (f.type, f.name))
-            export_acquisition.update_file(f.name, type=f.type)
-        if f.classification:
-            log.debug('Updating classification to %s for %s' % (f.classification, f.name))
-            export_acquisition.update_file_classification(f.name, f.classification)
-        if f.info:
-            log.debug('Updating info for %s' % (f.name))
-            export_acquisition.update_file_info(f.name, f.info)
-
-        export_acquisition.reload()
+                upload_file_path = os.path.join(temp_dir, f.name)
+                f.download(upload_file_path)
+    
+            # Upload the file to the export_acquisition
+            log.debug("Uploading %s to %s" % (f.name, export_acquisition.label))
+    
+            # Add logic around retrying failed uploads
+            max_attempts = 5
+            attempt = 0
+            while attempt < max_attempts:
+                attempt +=1
+                s = export_acquisition.upload_file(upload_file_path)
+                log.info('Upload status = {}'.format(s))
+                export_acquisition = fw.get_acquisition(export_acquisition.id)
+                file_names = [ x.name for x in export_acquisition.files ]
+                log.debug(file_names)
+                if os.path.basename(upload_file_path) not in file_names:
+                    log.warning('Upload failed for {} - retrying...'.format(os.path.basename(upload_file_path)))
+                else:
+                    log.info('Successfully exported: {}'.format(os.path.basename(upload_file_path)))
+                    break
+    
+    
+            # Update file metadata
+            if f.modality:
+                log.debug('Updating modality to %s for %s' % (f.modality, f.name))
+                export_acquisition.update_file(f.name, modality=f.modality)
+            if not f.modality and f.name.endswith('mriqc.qa.html'):
+                # Special case - mriqc output files do not have modality set, so
+                # we must set the modality prior to the classification to avoid errors.
+                export_acquisition.update_file(f.name, modality='MR')
+            if f.type:
+                log.debug('Updating type to %s for %s' % (f.type, f.name))
+                export_acquisition.update_file(f.name, type=f.type)
+            if f.classification:
+                log.debug('Updating classification to %s for %s' % (f.classification, f.name))
+                export_acquisition.update_file_classification(f.name, f.classification)
+            if f.info:
+                log.debug('Updating info for %s' % (f.name))
+                export_acquisition.update_file_info(f.name, f.info)
+    
+            export_acquisition.reload()
 
 
 def _cleanup(fw, creatio):
